requirements.txt
discord.py==1.4.1
pymongo==3.11.0
